Move data logic out of frontend API handlers
