+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/routing/frontend/WcifDataHandler.kt
@@ -6,11 +6,9 @@ import io.ktor.response.respond
 import io.ktor.routing.*
 import kotlinx.serialization.json.JsonNull
 import org.worldcubeassociation.tnoodle.server.RouteHandler
-import org.worldcubeassociation.tnoodle.server.plugins.EventPlugins
-import org.worldcubeassociation.tnoodle.server.webscrambles.Translate
+import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.WCIFCompetitorInfo
 import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.Competition
 import org.worldcubeassociation.tnoodle.server.webscrambles.zip.folder.PrintingFolder
-import java.util.*
 
 object WcifDataHandler : RouteHandler {
     override fun install(router: Route) {
@@ -19,16 +17,10 @@ object WcifDataHandler : RouteHandler {
                 post("competitors") {
                     val wcif = call.receive<Competition>()
 
-                    val competitorCountries = wcif.persons
-                        .map { it.countryIso2 }
-                        .map { it.isoString }
-                        .distinct()
+                    val detectedLocales = WCIFCompetitorInfo.detectTranslationLocales(wcif)
+                    val localeStrings = detectedLocales.map { it.toLanguageTag() }.sorted()
 
-                    val countryLocales = competitorCountries.flatMap { getTranslatedCountryLanguages(it) }
-                        .toSet() + Translate.DEFAULT_LOCALE
-
-                    val detectedLocales = countryLocales.distinct().map { it.toLanguageTag() }
-                    call.respond(detectedLocales)
+                    call.respond(localeStrings)
                 }
 
                 get("available") {
@@ -41,31 +33,15 @@ object WcifDataHandler : RouteHandler {
             post("best") {
                 val wcif = call.receive<Competition>()
 
-                val mbldResults = wcif.persons
-                    .flatMap { it.personalBests }
-                    .filter { it.eventPlugin == EventPlugins.THREE_MULTI_BLD }
-                    .mapNotNull { it.best.asMultiResult }
+                val bestMbldAttempt = WCIFCompetitorInfo.getBestMultiPB(wcif) { it.attempted }
 
-                val bestMbldAttempt = mbldResults.maxBy { it.attempted } ?: JsonNull
-                call.respond(bestMbldAttempt)
+                // ktor is not able to seralise null implicitly, only explicitly :/
+                val serialAttempt = bestMbldAttempt ?: JsonNull
+                call.respond(serialAttempt)
             }
         }
     }
 
-    private val COUNTRY_ISO2_LOCALES = Locale.getAvailableLocales()
-        .groupBy { it.country }
-
     private val AVAILABLE_LANGUAGE_TAGS
         get() = PrintingFolder.FMC_LOCALE_AVAILABLE_TAGS.toSortedSet()
-
-    private fun getTranslatedCountryLanguages(iso2: String): List<Locale> {
-        val candidateLocales = COUNTRY_ISO2_LOCALES[iso2] ?: return emptyList()
-
-        val translatedLocales = candidateLocales.filter { it in Translate.TRANSLATED_LOCALES }
-
-        val localLanguages = translatedLocales.filter { Locale(it.language, iso2) in Translate.TRANSLATED_LOCALES }
-        val generalLanguages = translatedLocales.flatMap { Translate.TRANSLATED_LOCALES.filter { l -> l.language == it.language } }
-
-        return localLanguages.takeIf { it.isNotEmpty() } ?: generalLanguages
-    }
 }
